+++ pkg/papi/edgehostname.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
+	"regexp"
 	"strings"
 
 	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v10/pkg/edgegriderr"
@@ -87,6 +88,14 @@ type (
 	CreateEdgeHostnameResponse struct {
 		EdgeHostnameLink string `json:"edgeHostnameLink"`
 		EdgeHostnameID   string `json:"-"`
+	}
+)
+
+var (
+	// domainPrefixPatterns maps domain suffixes to their respective regex patterns for validating domain prefixes
+	domainPrefixPatterns = map[string]*regexp.Regexp{
+		"akamaized.net": regexp.MustCompile(`^[A-Za-z]([A-Za-z0-9-]*[A-Za-z0-9])?$`),
+		"default":       regexp.MustCompile(`^[A-Za-z]([A-Za-z0-9.-]*[A-Za-z0-9])?(\.)?$`),
 	}
 )
 
@@ -254,6 +263,10 @@ func (p *papi) CreateEdgeHostname(ctx context.Context, r CreateEdgeHostnameReque
 		return nil, fmt.Errorf("%s: %w:\n%s", ErrCreateEdgeHostname, ErrStructValidation, err)
 	}
 
+	if err := validateDomainPrefix(r.EdgeHostname.DomainPrefix, r.EdgeHostname.DomainSuffix); err != nil {
+		return nil, err
+	}
+
 	logger := p.Log(ctx)
 	logger.Debug("CreateEdgeHostname")
 
@@ -286,4 +299,23 @@ func (p *papi) CreateEdgeHostname(ctx context.Context, r CreateEdgeHostnameReque
 	}
 	createResponse.EdgeHostnameID = id
 	return &createResponse, nil
+}
+
+func validateDomainPrefix(domainPrefix, domainSuffix string) error {
+	if len(domainPrefix) > 63 {
+		return fmt.Errorf("The edge hostname prefix must be 63 characters or less; you provided %d characters", len(domainPrefix))
+	}
+
+	pattern, exists := domainPrefixPatterns[domainSuffix]
+	if !exists {
+		pattern = domainPrefixPatterns["default"]
+	}
+
+	if !pattern.MatchString(domainPrefix) {
+		if domainSuffix == "akamaized.net" {
+			return fmt.Errorf("A prefix for the edge hostname with the \"akamaized.net\" suffix must begin with a letter, end with a letter or digit, and contain only letters, digits, and hyphens, for example, abc-def, or abc-123")
+		}
+		return fmt.Errorf("A prefix for the edge hostname with the \"%s\" suffix must begin with a letter, end with a letter, digit or dot, and contain only letters, digits, dots, and hyphens, for example, abc-def.123.456., or abc.123-def", domainSuffix)
+	}
+	return nil
 }