+++ CHANGELOG.md
@@ -16,6 +16,11 @@
 
 
 
+* PAPI
+  * Added name format validations and a size limit for the Domain Prefix. 
+
+
+